backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-google-genai
chromadb
pydantic
requests